releases: Remove visual clutter and unneeded doc links
For many versions there was listed in the documentation section a link
"API" that linked to a snapshot of the headers on gitweb.
Now that the manpages are published on the website, users should just
refer to that.

For many versions there was also list a link to build instructions. Remove
that too.

The final result is a much cleaner page that presents exactly the
information users need.